[step3] test: Driver.kt 생성, drive 호출, drive 가능 기준판단 함수를 위임함
