+++ src/test/kotlin/step3/racingcar/CarTest.kt
@@ -6,18 +6,6 @@ import step3.racingcar.car.Car
 import step3.racingcar.car.Position
 
 class CarTest {
-
-    @Test
-    fun `4 이상 숫자가 나오면 움직인다`() {
-        val moveNumber = 5
-        val nonMoveNumber = 3
-        val boundaryNumber = 4
-
-        assertThat(Car().isMovable(moveNumber)).isTrue
-        assertThat(Car().isMovable(nonMoveNumber)).isFalse
-        assertThat(Car().isMovable(boundaryNumber)).isTrue
-    }
-
     @Test
     fun `차 이동&멈춤의 상태를 저장한다`() {
         // given
